+++ allensdk/__init__.py
@@ -35,7 +35,7 @@
 #
 import logging
 
-__version__ = '2.13.5'
+__version__ = '2.13.9'
 
 
 try: