+++ allensdk/brain_observatory/ecephys/current_source_density/__main__.py
@@ -11,24 +11,34 @@ from pathlib import Path
 
 from typing import Optional
 
-from ._schemas import InputParameters, OutputParameters
-from ._current_source_density import (
-    accumulate_lfp_data,
-    compute_csd,
-    extract_trial_windows
-)
-from ._filter_utils import filter_lfp_channels, select_good_channels
-from ._interpolation_utils import (
-    interp_channel_locs,
-    make_actual_channel_locations,
-    make_interp_channel_locations
-)
+from scipy.spatial.qhull import QhullError
+
+from allensdk.brain_observatory.ecephys.current_source_density._schemas \
+    import \
+    InputParameters, OutputParameters
+from allensdk.brain_observatory.ecephys.current_source_density.\
+    _current_source_density import (
+        accumulate_lfp_data,
+        compute_csd,
+        extract_trial_windows
+    )
+from allensdk.brain_observatory.ecephys.current_source_density._filter_utils \
+    import filter_lfp_channels, select_good_channels
+from allensdk.brain_observatory.ecephys.current_source_density\
+    ._interpolation_utils import (
+        interp_channel_locs,
+        make_actual_channel_locations,
+        make_interp_channel_locations
+    )
 from allensdk.brain_observatory.ecephys.file_io.continuous_file import (
     ContinuousFile
 )
 from allensdk.brain_observatory.argschema_utilities import (
     write_or_print_outputs, optional_lims_inputs
 )
+
+from allensdk.brain_observatory.ecephys.lfp_subsampling.subsampling \
+    import remove_lfp_noise
 
 
 def get_inputs_from_lims(args) -> dict:
@@ -75,6 +85,13 @@ def run_csd(args: dict) -> dict:
 
     stimulus_table = pd.read_csv(args['stimulus']['stimulus_table_path'])
 
+    # backwards compatibility
+    stimulus_table['stimulus_name'] = stimulus_table['stimulus_name'].apply(
+        lambda x: args['stimulus']['key'] if x == 'flashes' else x)
+    if args['start_field'] not in stimulus_table:
+        stimulus_table = stimulus_table.rename(
+            columns={'Start': args['start_field']})
+
     probewise_outputs = []
     for probe_idx, probe in enumerate(args['probes']):
         logging.info('Processing probe: {} (index: {})'.format(probe['name'],
@@ -109,10 +126,17 @@ def run_csd(args: dict) -> dict:
         else:
             lfp_channels = np.arange(0, probe['total_channels'])
 
+        lfp_referenced = remove_lfp_noise(
+            lfp=lfp_raw,
+            surface_channel=probe['surface_channel'],
+            channel_numbers=lfp_channels,
+            max_out_of_brain_channels=args['max_out_of_brain_channels']
+        )
+
         logging.info('Accumulating LFP data')
         accumulated_lfp_data = accumulate_lfp_data(
             timestamps=timestamps,
-            lfp_raw=lfp_raw,
+            lfp_raw=lfp_referenced,
             lfp_channels=lfp_channels,
             trial_windows=trial_windows,
             volts_per_bit=args['volts_per_bit']
@@ -141,11 +165,31 @@ def run_csd(args: dict) -> dict:
             0,
             accumulated_lfp_data.shape[1]
         )
-        interp_lfp, spacing = interp_channel_locs(
-            lfp=filt_lfp,
-            actual_locs=clean_actual_locs,
-            interp_locs=interp_locs
-        )
+        if len(clean_channels) == 0:
+            logging.error(f'There are no clean channels. Skipping probe '
+                          f'{probe["name"]}')
+            probewise_outputs.append({
+                'name': probe['name'],
+                'csd_path': None,
+                'clean_channels': clean_channels.tolist()
+            })
+            continue
+        try:
+            interp_lfp, spacing = interp_channel_locs(
+                lfp=filt_lfp,
+                actual_locs=clean_actual_locs,
+                interp_locs=interp_locs
+            )
+        except QhullError:
+            logging.error(f'There are only {len(clean_channels)} '
+                          f'clean channels, which is not enough for '
+                          f'interpolation. Skipping probe {probe["name"]}')
+            probewise_outputs.append({
+                'name': probe['name'],
+                'csd_path': None,
+                'clean_channels': clean_channels.tolist()
+            })
+            continue
 
         logging.info('Averaging LFPs over trials')
         trial_mean_lfp = np.nanmean(interp_lfp, axis=0)
@@ -171,6 +215,7 @@ def run_csd(args: dict) -> dict:
         probewise_outputs.append({
             'name': probe['name'],
             'csd_path': probe['csd_output_path'],
+            'clean_channels': clean_channels.tolist()
         })
 
     return {
@@ -200,7 +245,8 @@ def write_csd_to_h5(path: Path, csd: np.ndarray, relative_window,
 def main():
 
     logging.basicConfig(format=('%(asctime)s:%(funcName)s'
-                                ':%(levelname)s:%(message)s'))
+                                ':%(levelname)s:%(message)s'),
+                        level=logging.INFO)
     parser = optional_lims_inputs(sys.argv, InputParameters,
                                   OutputParameters, get_inputs_from_lims)
     output = run_csd(parser.args)