+++ allensdk/test/brain_observatory/ecephys/test_ecephys_project_lims_api.py
@@ -47,10 +47,15 @@ class MockSelector:
         {"session_ids": [1, 2, 3]},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "filters_sessions":
                 lambda st:
                     re.compile(r".+and es.id in \(1,2,3\).*",
                                re.DOTALL).match(st) is not None
+=======
+            "filters_sessions": lambda st: re.compile(
+                r".+and es.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -62,10 +67,15 @@ class MockSelector:
         {"unit_ids": [1, 2, 3]},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "filters_units":
                 lambda st:
                     re.compile(r".+and eu.id in \(1,2,3\).*",
                                re.DOTALL).match(st) is not None
+=======
+            "filters_units": lambda st: re.compile(
+                r".+and eu.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -77,6 +87,7 @@ class MockSelector:
         {"channel_ids": [1, 2, 3], "probe_ids": [4, 5, 6]},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "filters_channels":
                 lambda st:
                     re.compile(r".+and ec.id in \(1,2,3\).*",
@@ -85,6 +96,13 @@ class MockSelector:
                 lambda st:
                     re.compile(r".+and ep.id in \(4,5,6\).*",
                                re.DOTALL).match(st) is not None
+=======
+            "filters_channels": lambda st: re.compile(
+                r".+and ec.id in \(1,2,3\).*", re.DOTALL).match(
+                st) is not None,
+            "filters_probes": lambda st: re.compile(
+                r".+and ep.id in \(4,5,6\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -96,6 +114,7 @@ class MockSelector:
         {"published_at": "2019-10-22"},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "checks_pa_not_null":
                 lambda st:
                     re.compile(r".+and es.published_at is not null.*",
@@ -104,6 +123,14 @@ class MockSelector:
                 lambda st:
                     re.compile(r".+and es.published_at <= '2019-10-22'.*",
                                re.DOTALL).match(st) is not None
+=======
+            "checks_pa_not_null": lambda st: re.compile(
+                r".+and es.published_at is not null.*", re.DOTALL).match(
+                st) is not None,
+            "checks_pa": lambda st: re.compile(
+                r".+and es.published_at <= '2019-10-22'.*", re.DOTALL).match(
+                st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -115,6 +142,7 @@ class MockSelector:
         {"published_at": "2019-10-22", "session_ids": [1, 2, 3]},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "checks_pa_not_null":
                 lambda st:
                     re.compile(r".+and es.published_at is not null.*",
@@ -127,6 +155,16 @@ class MockSelector:
                 lambda st:
                     re.compile(r".+and es.id in \(1,2,3\).*",
                                re.DOTALL).match(st) is not None
+=======
+            "checks_pa_not_null": lambda st: re.compile(
+                r".+and es.published_at is not null.*", re.DOTALL).match(
+                st) is not None,
+            "checks_pa": lambda st: re.compile(
+                r".+and es.published_at <= '2019-10-22'.*", re.DOTALL).match(
+                st) is not None,
+            "filters_sessions": lambda st: re.compile(
+                r".+and es.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -138,6 +176,7 @@ class MockSelector:
         {"published_at": "2019-10-22", "session_ids": [1, 2, 3]},
         pd.DataFrame({"id": [5, 6], "something": [12, 14]}),
         {
+<<<<<<< HEAD
             "checks_pa_not_null":
                 lambda st:
                     re.compile(r".+and es.published_at is not null.*",
@@ -150,6 +189,16 @@ class MockSelector:
                 lambda st:
                     re.compile(r".+and es.id in \(1,2,3\).*",
                                re.DOTALL).match(st) is not None
+=======
+            "checks_pa_not_null": lambda st: re.compile(
+                r".+and es.published_at is not null.*", re.DOTALL).match(
+                st) is not None,
+            "checks_pa": lambda st: re.compile(
+                r".+and es.published_at <= '2019-10-22'.*", re.DOTALL).match(
+                st) is not None,
+            "filters_sessions": lambda st: re.compile(
+                r".+and es.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14]},
@@ -159,6 +208,7 @@ class MockSelector:
     [
         "get_sessions",
         {"published_at": "2019-10-22", "session_ids": [1, 2, 3]},
+<<<<<<< HEAD
         pd.DataFrame({"id": [5, 6],
                       "something": [12, 14],
                       "genotype": ["foo", np.nan]}),
@@ -174,6 +224,19 @@ class MockSelector:
             "filters_sessions":
                 lambda st: re.compile(r".+and es.id in \(1,2,3\).*",
                                       re.DOTALL).match(st) is not None
+=======
+        pd.DataFrame({"id": [5, 6], "something": [12, 14],
+                      "genotype": ["foo", np.nan]}),
+        {
+            "checks_pa_not_null": lambda st: re.compile(
+                r".+and es.published_at is not null.*", re.DOTALL).match(
+                st) is not None,
+            "checks_pa": lambda st: re.compile(
+                r".+and es.published_at <= '2019-10-22'.*", re.DOTALL).match(
+                st) is not None,
+            "filters_sessions": lambda st: re.compile(
+                r".+and es.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"something": [12, 14], "genotype": ["foo", "wt"]},
@@ -183,6 +246,7 @@ class MockSelector:
     [
         "get_unit_analysis_metrics",
         {"ecephys_session_ids": [1, 2, 3]},
+<<<<<<< HEAD
         pd.DataFrame({"id": [5, 6],
                       "data": [{"a": 1, "b": 2}, {"a": 3, "b": 4}],
                       "ecephys_unit_id": [10, 11]}),
@@ -190,6 +254,14 @@ class MockSelector:
             "filters_sessions":
                 lambda st: re.compile(r".+and es.id in \(1,2,3\).*",
                                       re.DOTALL).match(st) is not None
+=======
+        pd.DataFrame(
+            {"id": [5, 6], "data": [{"a": 1, "b": 2}, {"a": 3, "b": 4}],
+             "ecephys_unit_id": [10, 11]}),
+        {
+            "filters_sessions": lambda st: re.compile(
+                r".+and es.id in \(1,2,3\).*", re.DOTALL).match(st) is not None
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
         },
         pd.DataFrame(
             {"id": [5, 6], "a": [1, 3], "b": [2, 4]},
@@ -198,11 +270,11 @@ class MockSelector:
     ]
 ])
 def test_pg_query(method_name, kwargs, response, checks, expected):
-
     selector = MockSelector(checks, response)
 
     with mock.patch("allensdk.internal.api.psycopg2_select",
                     new=selector) as ptc:
+<<<<<<< HEAD
 
         api = epla.EcephysProjectLimsApi.default(
                 lims_credentials=mock_lims_credentials)
@@ -214,6 +286,13 @@ def test_pg_query(method_name, kwargs, response, checks, expected):
             obtained,
             check_like=True,
             check_dtype=False)
+=======
+        api = epla.EcephysProjectLimsApi.default(
+            lims_credentials=mock_lims_credentials)
+        obtained = getattr(api, method_name)(**kwargs)
+        pd.testing.assert_frame_equal(expected, obtained, check_like=True,
+                                      check_dtype=False)
+>>>>>>> 0f4ac934ef97011ad3ea56eda63cf8448c4d8eaa
 
         any_checks_failed = False
         for name, result in ptc.passed.items():
@@ -280,7 +359,6 @@ class MockHttpEngine:
     ]
 ])
 def test_file_getter(method, kwargs, query_pattern, pg_engine_cls):
-
     api = epla.EcephysProjectLimsApi(
         postgres_engine=pg_engine_cls(query_pattern),
         app_engine=MockHttpEngine()